Moved Dishonoured and Borderlands pointers into game event file.
