--- Project-Aurora/Project-Aurora/Profiles/Borderlands 2/Borderlands2Pointers.cs
@@ -1,12 +1,0 @@
-﻿using Aurora.Utils;
-
-namespace Aurora.Profiles.Borderlands2
-{
-    public class Borderlands2Pointers
-    {
-        public PointerData Health_maximum;
-        public PointerData Health_current;
-        public PointerData Shield_maximum;
-        public PointerData Shield_current;
-    }
-}
