--- Project-Aurora/Project-Aurora/Profiles/Dishonored/DishonoredPointers.cs
@@ -1,14 +1,0 @@
-﻿using Aurora.Utils;
-
-namespace Aurora.Profiles.Dishonored
-{
-    public class DishonoredPointers
-    {
-        public PointerData ManaPots;
-        public PointerData HealthPots;
-        public PointerData CurrentHealth;
-        public PointerData MaximumHealth;
-        public PointerData CurrentMana;
-        public PointerData MaximumMana;
-    }
-}
