+++ Project-Aurora/Project-Aurora/Profiles/Dishonored/GameEvent_Dishonored.cs
@@ -28,4 +28,14 @@ namespace Aurora.Profiles.Dishonored
             gameState.Player.HealthPots = reader.ReadInt(pointers.HealthPots);
         }
     }
+
+    public class DishonoredPointers
+    {
+        public PointerData ManaPots;
+        public PointerData HealthPots;
+        public PointerData CurrentHealth;
+        public PointerData MaximumHealth;
+        public PointerData CurrentMana;
+        public PointerData MaximumMana;
+    }
 }
